requirements.txt
opencv-python==4.11.0.86
opencv-contrib-python==4.11.0.86
numpy==1.26.4
Pillow==11.2.1
pandas==2.2.3
openpyxl==3.1.5
mediapipe==0.10.21

#python==3.11